src/requirements.txt
requests 
openpyxl 
python-dotenv 
pandas
beautifulsoup4
selenium 
webdriver-manager 
tqdm
